chore: enhance false-positive visual test reports
The retry feature we lately added seems to work as expected.

But when there are re-try attempts, our jest-screenshot dependency creates a report. This report gets not removed and during the global teardown event, we then commit the reports to a new branch.

With this commit, we check put the name of a test in a cache, and remove it on success. At the end we then have hopefully an empty list, and will not commit the reports.