chore(CI): add CACHE_VERSION
With this fix, we can quickly invalidate all caches by changing the ENV in GitHub env settings. This is because, there is no way to invalidate the GitHub Actions cache. So this is a recommended solution.

Its only meant to use for special cases. Until now, we never did really need that. But when we need it, we need to do it quickly (most probably).