chore(repo): declare deps correctly towards Yarn PnP
This PR only fixes missing, wrong configured deps when using PnP

Prettier (and ESLint) in VSCode does still not work with PnP

Therefore, we still may have to wait.

Why is PnP interesting? Because of the "zero" install on CI (and locally) – but most for our visual tests on CI, because the macOS form GitHub Actions are very slow on that step.