fix(Breadcrumb): make animation work when declarative declared
When a Breadcrumb was declared like so:

```jsx
    <Breadcrumb>
      <Breadcrumb.Item
        text="Page item 1"
      />
      <Breadcrumb.Item
        text="Page item 2"
      />
    </Breadcrumb>
```

The animation did not work nicely then. Why? We do not know the count of each item.

[CSB](https://codesandbox.io/s/eufemia-breadcrumb-animation-issue-thsgsm?file=/src/App.tsx)

This PR ensures we always know the number of each item during the each render phase, based on the text content – because that's the only reference we have, that will not change, during open/close.

Because I wanted to share `CurrentVariant` type, I ended up moving all types to a new  `/types.ts` file. Sorry for the noise, it should have been in a separate PR.