chore(Release): make "next" tag to a pre-release (#1692)
Something is not quite right. We follow the docs from semantic-release with the existing setup, but still, the next version got released under NPM next dist-tag, but without a suffix of `-next` – and also, the git tag did not get it.

So, for now, I think, we should mark it as a pre-release instead and find a way of not publish any release notes in the GitHub release section. But this has to be done in a separate PR.

Also, running the verify tests before release, is not optimal with the current setup, because it runs regardless, and not in band. So I think we want to remove `workflow_run`. And rather include them in the jobs directly.