chore(Portal): add scroll-padding-top to compensate for sticky header
Now that we do not need to support IE, we can use scroll-padding-top to compensate for our sticky header.

This is useful, when a location hash is used to scroll to a certain element with an ID. Does not effect our heading hash handling, there we already deal with a offset (padding).

Also, we do not need (neither should) to set a fixed height to the header. The given padding should be enough.