chore(CI): API rate limit exceeded for installation ID 15060380
This is an issue we can't do much about it. It looks like the GitHub Action we use, uses an GitHub API for checking if its ready to go. But there is a rate limit. Its a [known issue](h).

For nor we can try to increate the interval from 3 sec to 10 sec.