chore(Portal): load the DNB UI theme as embedded CSS
This effects only the prod build. From what I'v seen during the period we had this enabled– loading the needed CSS via a file, delays other resources and gives us a much worse UX. For sure, only on the first couple of seconds. So I suggest to rather load a little more CSS, when switching the theme.