chore(VisualTests): remove wrapper after taking screenshot (#2474)
Based on findings from PR #2348 we need to remove the added wrapper
after a visual test has run.

**Background:**

Because, when having stacked several testing targets, they can/will
interfere the width of each other. So, when running one at a time
locally, the snapshot will not match in comparison to when run all
together.

The same negative effect we saw on running a dev build in comparison to
running the tests on the local dev setup. There I'm not 100% why. But I
guess, it has to do with the existing HTML and re-hydration.