chore(CI): fix postbuild run in e2e action (#2515)
Ensure we actually run the "postbuild" step on the main (and e.g. v10)
branch only. This corrects the code as it was intended to be, because we
need to check against a string, and not a boolean.

With that change, we should safe about 2.5 min on normal PR branches.

Also, we run the "postbuild" step in the **Verify** action anyway.