fix(Forms): sets SubHeading's size to medium (#2581)
## Summary

<!-- Explain the **motivation** for making this change. What existing
problem does the pull request solve? -->

I experienced in an attempt to implement using the SubHeading component
that the size of the text was dependent on outside components. The
hx-number might need to vary based on surroundings, but the point of
having `MainHeading` and `SubHeading` in Eufemia Forms was that it would
be predictable what size the text became. For that reason, I set the
SubHeading hard coded to "medium" similar to how MainHeading has been
coded to "large", and added a few more examples to trigger actual
results of these types of combinations for testing purposes.

copilot:summary

## Details

copilot:walkthrough

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran
and their output, screenshots / videos if the pull request changes UI.
-->