fix(Forms): precede children over title in Selection (#2664)
If a `<Field.Option>` in a `dropdown` `<Field.Selection>` has `title`
and `children` properties, the `title` property precedes the `children`
property. That is reversed logic from the `<Field.Option>` component,
where the `children` property precedes the `title` property.

This commit aligns the logic with the `<Field.Option>` component, making
the `children` property precede the `title` property.