feat(Section): add responsive props: breakout, outline, cornerRadius, backgroundColor, textColor and innerSpace (#2791)
The origin idea to deliver a solution for handling responsive designs
with a more declarative approach comes from @henit – thank you so much!

Inspired by the idea, this PR enhances the Section component with a
couple more props. All of them do also accept an object with media query
breakpoints, so their values can dynamically respond to a "screen size".

- `breakout={boolean}` or e.g. `breakout={{ small: boolean }}`
- `roundedCorner={boolean}` or e.g. `roundedCorner={{ small: boolean }}`
- `outline={boolean|string}` or e.g. `outline={{ small: 'black' }}`
- `backgroundColor={string}` or e.g. `backgroundColor={{ small: 'white'
}}`
- `textColor={string}` or e.g. `textColor={{ small: 'black-80' }}`


While inner spacing support is added by the new space component feature:

- `innerSpace={string}` or e.g. `innerSpace={{ small: { top: 'small' }
}}`


All of the media query logic happens in CSS, so it will serve SSR apps
very well.

This PR should not change the original functionality of the Section
component. It only enhances it with new features. Even a lot of the
logic got rewritten in order to utilize CSS custom properties in a good
way.