feat: add "use client" to components that use client side only React APIs (support for RSC) (#2874)
We have received several request to add a ["use
client"](https://react.dev/reference/react/use-client) directive on top
of each React component file that uses client side only React APIs.

The reason is because Next.js takes in use a beta version of React to
make RSC work (which is not OK in my opinion – its a super aggressive
move 😣).

To make a first move on this from our side, I think we should automate
this. And one way to do so, is to use Babel.js – because we use it in
our tool-chain.

There is already a very well written and tested plugin, written by Kyle
Tsang I suggest to try out.

👉
[babel-plugin-transform-next-use-client](https://github.com/kyletsang/babel-plugin-transform-next-use-client)

This must be the first time we take in use a dependency with 0 GitHub
stars. If it turns out to work as expected, we should all give it a star
of course!

**Other notes:**

- I have tested it locally and also added a post build test.
- We skip rollup builds (UMD/ESM) bundles. But they are not uses in
Next.js anyway.
- Time will tell of we need to [add more custom
config](https://github.com/kyletsang/babel-plugin-transform-next-use-client#react-client-apis-in-custom-modules).