fix(PhoneNumber): keep selected countryCode value on blur (#2869)
The change: When the user changes the countryCode value – without making
a "selection" – and then the input blurs, it will show the current
selected value. This way, it will not be possible to "clear" a country
code or add a custom.

But that also requires us to deliver all possible available codes. The
error rate will be way lower for sure. Now, its more like a Dropdown,
but with fee text search ability.

Here's [a
demo](https://eufemia-git-feat-phone-number-keep-value-eufemia.vercel.app/uilib/extensions/forms/feature-fields/PhoneNumber/demos/).