feat(Field.Number): add step control functionality (#3140)
Adds a new prop `showStepControls` to the Number component. When `true`,
this prop enables control buttons that adhere to a `step` property that
is also added, as well as taking min/max values into account.

This variant of the component is Sbanken-ready, and a number of tests
and screenshot tests has been implemented in relation to it.