feat(NumberUtils.format): Only return object if returnAria: true
The function `format` only returns an object if the option `returnAria` was passed in as `true`, but TypeScript doesn’t know this.

With this change, we’re making an overload for the function that only returns the object in the right situation. This way, the caller doesn’t need to check the returned value before being able to use it as a string (or number).