feat(FieldBlock): enhance support for composition fields
This commit adds docs, examples and tests on how to create composition fields. It also adds easy handling of the hasError prop we get from useDataValue.

We also lists combined status messages with a list, instead of a pipe. And if a status message appears to be the same, we now show just one of them.

Update demos.mdx

Co-authored-by: Joakim Bjerknes <joakbjerk@gmail.com>