chore: add DocTable for easy documenting properties inside components directory and replace input props with this approach
This PR introduces a new way of how we may/can place and write our component properties. The idea is that we move it from Markdown Tables to TypeScript, but inside the component directory itself.

This opens up for more flexibility and much better Git diffs as we never will change the whole table, just because the max lengths has changed of one description.