fix(HeightAnimation): ensure Radio buttons persist their state (#3535)
Why it effects radio buttons: The `name` attribute is used to create a
"group". But when a new element gets added, that has the same `name`, it
conflicts with the existing group, and removes the checked state of the
current checked radio button, as there is a second "checked" one.