feat(Forms): deprecate filterSubmitData in Form.Handler and return `filterData` in onSubmit and onChange instead (#3873)
The property `filterSubmitData` on the `Form.Handler` was a way to
define a filter, so data returned in the `onSubmit` got filtered.

But I think we rather should return the filter function and let devs do
it right where it happens. We do that already in `useData` and in
`getData` that way. And now we also return `filterData` in the
`onChange` event.

The motivation is to keep consistency and rather offer declarative ways
of defining filters.

---------

Co-authored-by: Anders <anderslangseth@gmail.com>