chore(Forms): Behaviour => Behavior (#3903)
"Behaviour" is the British English spelling.
"Behavior" is the American English spelling.

Our code is written in en-us English, while the translation English is
en-gb.

There are a couple more. One in the Accordion component
(`expandBehaviour`) as a property. But this I think we should keep, as
it needs to be deprecated. Feel free to make another PR 🙏