fix(Forms): ensure labels do update when they change (async fields) (#3910)
In the submit indicator component we check if a label/text did wrap to a
new line, because of too little space. In this check routine we
insert/manipulate the DOM via `element.textContent`. But this check
messes up with React and makes so that the given children is
re-inserted, so new children/label will not be shown.
We can fix this chancing the `key` of the element.