fix(Forms): remove unwanted Iterate.EditContainer error confirmation when using required on Form.Handler (#3938)
When `required` was set on the Form.Handler, this code;

```tsx
<Form.Handler required>
  <Iterate.Array path="/myList">
    <Iterate.ViewContainer>ViewContainer</Iterate.ViewContainer>
    <Iterate.EditContainer>EditContainer</Iterate.EditContainer>
  </Iterate.Array>
  <Iterate.PushButton path="/myList" pushValue={{}} />
</Form.Handler>
```

did result in this;

<img width="318" alt="Screenshot 2024-09-15 at 20 20 06"
src="https://github.com/user-attachments/assets/f9d872e3-7fa4-4c5e-8ceb-d27a921406fe">

... when pressing the done button.

This happened because the remove button did use the hook `useFieldProps`
which is meant to be used by fields. So it suddenly did validate as
required.