feat(Forms): add `inheritVisibility` to each `Value.*` component (#3948)
This PR adds a new property called `inheritVisibility`. When it is used,
it will look for a field with the same path. If it exists, it will sync
its visibility to the mounting state of the field.

A following up PR will make the same, but for the submit data.

- [x] We should test this on the ChildrenWithAge block

---------

Co-authored-by: Joakim Bjerknes <joakbjerk@gmail.com>
Co-authored-by: Anders <anderslangseth@gmail.com>