chore(Forms): fix initial `transformOut` call
In PR #3975 (not released at this point of time), we added a new `transformOut` function to the `useFieldProps` hook. This function is with that also called when the field is initialized (and there is a path and Form.Handler). It is used to transform the value before it is stored in the data context. However, this function, when changing the value, it updates the data context, but it should not update the internal value.

This out of sync situation could lead to an infinite loop.

By double checking if the data context value did change, we can avoid this infinite loop.

The provided test would fail without the fix.