chore(Forms): fix initial `transformOut` call (#3980)
In PR #3975 (not released at this point of time), we added a new
`transformOut` function to the `useFieldProps` hook. This function is
with that also called when the field is initialized (and there is a path
and Form.Handler). It can be used to transform the value before it is
stored in the data context. However, this function, when changing the
value in the context, it will re-render and with that, try to do the
same again, but we had no check for this case, so it can/will result in
an infinite loop.

By double checking if the data context value did change since last time,
we can avoid this infinite loop.

The provided tests would fail with an infinite loop without the fix.