feat(Forms): implement TypeScript version of `json-pointer` (#4000)
I believe this is the right step for now because the dependency we’re
using is written in CJS and doesn't treeshake unused parts. Since we
rely on this pointer in the default Eufemia Context, reducing its
footprint is my main motivation.

By using the native `foreach` method and the export approach in this PR,
we can ship smaller bundle (maybe 20%). In the future, we may consider
submitting a PR to the original dependency to address this issue
directly.

PR no. 4000