fix(forms): add `sessionStorageId` support to Field.Upload with empty file list rendering (#4339)
When using `sessionStorageId` on `Form.Handler` alongside `Field.Upload`
with a specified path, file data is stored. However, during
serialization, Blob information is not preserved in session storage. As
a result, reading this data causes `Field.Upload` to throw an exception.

This PR ensures invalid files are not rendered, preventing such errors.

In the future, we could explore storing Blobs in IndexedDB, but that
would require a more extensive effort.