fix(HeightAnimation): ensure height does not lock during animation when nested (#4493)
The animation is slowed down to 4s in order to see the issue. 
Keep an eye on the container edge, just below the "Velg land" field.
The outer height animation sets a "fixed" height, even only the inner
content height gets updated. The fix is to tell all "outer" height
animations to not animated (keep the auto height), when an "inner"
height gets changed.

Before:




https://github.com/user-attachments/assets/5c53d017-c344-4a4f-8179-75908df7e3aa


After:



https://github.com/user-attachments/assets/84229310-1b39-4790-b5d4-a7c70ba1f6da