feat(Forms): show field indicator inline without shifting content below (#4553)
This is only a slight adjustment, not meant to be the final solution for
our FieldBlock indicator. The reason for this adjustment is that having
it "inline" by pushing other content down, it creates a very distracting
UX. The downside? It now can overlap like in the screenshot below:

<img width="502" alt="Screenshot 2025-02-11 at 08 23 31"
src="https://github.com/user-attachments/assets/73d7ea56-08af-40c9-88e3-220e19d901e6"
/>