removes wording with current and currently (#4577)
In the following sentences of this PR, I don't feel it make sense to use
the word current/currently.
As "everything" is sort of currently/now, so I'm not sure what value it
adds to use the word in our docs. It could make sense to say how it is
currently if we also add info about how we envision it in the future or
so, but that we don't do in these sentences.