feat: use @dnb/browserslist-config for when creating the bundle (#4704)
My research so far is:

There is a difference in the output. But its minimal. Basically, every
file (over 400) have some more imports such as this:

```js
import "core-js/modules/web.dom-collections.iterator.js";
```

Also, we may add a list of the supported browsers in a human readable
table/list. Here is a PR that would make that possible:
https://github.com/dnbexperience/browserslist-config

---------

Co-authored-by: Anders <anderslangseth@gmail.com>