feat(GlobalError): deprecate `status` in favor of `statusCode` (#4819)
I like the word `statusCode` better than `status`.
It also seems that other sources do as well:
- https://en.wikipedia.org/wiki/List_of_HTTP_status_codes
- https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status

Changing from `status` to `statusCode` also helps with reducing the
confusion with the property `code`, see docs
https://eufemia.dnb.no/uilib/components/global-error/properties/

But I'll also open a PR to change that property name as well, to reduce
the confusion.

Motivation:
https://dnb-it.slack.com/archives/CMXABCHEY/p1742973159203409