fix(Heading): warning when increasing heading to be `h2` when previous sibling is `h1` (#4914)
Motivation:
https://dnb-it.slack.com/archives/CMXABCHEY/p1743769418111539
CSB: https://codesandbox.io/p/devbox/value-padding-forked-5lrp2k

I do not think we should warn about this, as I believe:
```
<Heading>heading</Heading>
<Heading increase>subheading</Heading>
``` 
should result in h1 and h2 which it does, but without warnings.

I do think the reason why we log the warning today, is because we
automatically change the second heading to be a h2 in the snippet:
```
<Heading>heading</Heading>
<Heading>subheading</Heading>
``` 
So hence, our code tries/thinks the following code will increase the
heading level to h3 of the second heading which is already made into a
h3:

```
<Heading>heading</Heading>
<Heading increase>subheading</Heading>
``` 


I think the code perhaps could be simplified or made a tiny bit easier
to read.

---------

Co-authored-by: Tobias Høegh <tobias@tujo.no>