chore: add `Eufemia.versions` support (#5234)
Also, it should with that be possible to get a short sha. We may
document that as a feature if we see that it works.

Here is a test setup: https://stackblitz.com/edit/dsuwq5hy

<img width="208" alt="Screenshot 2025-06-05 at 14 00 40"
src="https://github.com/user-attachments/assets/37e8e7db-2f43-468b-80a3-36dfc5b83ca9"
/>