chore: update `properties.js` files (#5270)
When running a local build, they get update. Also, during a release they
will be updated and a part of the release anyway. Its just, when running
it locally we then can commit the auto generated files. But we could in
future "fail" if auto generated stiff is generated, without being
committed.