feat(Forms): add address suggestions and autofill to the Bring API connector (#4575)
Addition to #4554
PR preview:
https://eufemia-git-feat-forms-api-connectors-bring-addr-1a098a-eufemia.vercel.app/uilib/extensions/forms/Connectors/Bring/#address-suggestions-api


https://github.com/user-attachments/assets/75babeeb-2e7f-4170-aee1-4d4fa7af36c5

There are surely many ways of how to handle such a suggestions
connector.
The goal was to use an Autocomplete instead of a string input. That's
the reason I ended up with a `element` replacement.

```tsx
<Field.Address.Street
  element={addressSuggestionsElement}
/>
```

Here is how its created:

```ts
const addressSuggestionsElement = withConfig(
  Connectors.Bring.address.suggestionsElement,
  {
    cityPath: '/city',
    postalCodePath: '/postalCode',
  }
)
```

That's the reason why this PR also creates a `Field.Autocomplete` so we
easily can take advantage of the `useFieldProps` hook and get access to
`handlePathChangeUnvalidated` to update other fields with the selected
data.

How ever, there is more work to do:

- [x] Enhance typings.
- [x] Add tests.
- [x] Add documentation.

---------

Co-authored-by: Anders <anderslangseth@gmail.com>