fix(Field.PhoneNumber): truncate to 8 digits when changing country code to `NO` (#5266)
Motivation:

https://dnb.enterprise.slack.com/archives/CMXABCHEY/p1750159051952929

I find it a bit scary to overwrite/change the input of the user. 
Perhaps an even better solution would be to still display the user
entered value together with some error/warning message or so?
Not 100% sure how to do this while still using the mask and defaultMask.