chore(Field.Selection): refactor tests to use userEvent (#5279)
I just have seen some of the tests fail sometimes. But I don't think
this will help much. Anyway, it kind of feels more right to use
`userEvent` as we the also are able to "simulate" a delayed even
handling in order to debug:

```ts
const userEvent = original.setup({ delay: 100 })
```