fix(Input): ensure error, focus, and hover states work with autofilled content (#5424)
Fixes #5217

I also tried adding a visual test for these styles. There is a way to do
this via Chromium and its DevTools, but it would require a significant
refactoring.
Another option would be to use a "simulation" class, but I think this
approach is too fragile from a maintenance perspective and could
negatively impact CSS specificity in certain cases.