fix(Upload): call `setInternalFiles` on `setFiles` (#5571)
Motivation:
https://dnb-it.slack.com/archives/CMXABCHEY/p1759744444493039 when
clearing files using `setFiles`.



As Upload use `internalFiles` to determine if it should show the "Det er
begrenset hvor mange filer du kan laste opp"-error message, I believe we
have to update change this when using `setFiles` , see code:
https://github.com/dnbexperience/eufemia/blob/d83184d070b9622ecfd998e9394bc6210e9da2f3/packages/dnb-eufemia/src/components/upload/UploadStatus.tsx#L11


This PR does not reset the shared state / context on unmount when an
`id` is provided. Not entirely sure if or how to do so, or if we want to
do so, wouldn't that be potentially breaking? As the whole point of
using an `id` and shared state / context is that _The useUpload hook can
be placed in any location in your application, and does not need to be
where the Upload component is used._ -
https://eufemia.dnb.no/uilib/components/upload/#how-to-use-the-upload-component
Or it possible to achieve both?