fix(Dropdown, Autocomplete): keyboard navigation fixes (#5444)
Commit message:

--- 

Changes:
- Keyboard navigation no longer focuses the container `ul` between
loops.
- `aria-activedescendant` now controlled by the `DrawerList` context.
- If an item is selected, it receives focus when the list is opened,
instead of the previously focused element.

Fixes:
- Keyboard navigation no longer loops to the selected item.
- `Autocomplete` now has a `aria-activedescendant` set when first opened
so screen readers can navigate the list
- `Autocomplete` now loops in both directions, not just from the bottom.
- Pressing any key no longer resets focus unless it's a search with a
result.

--- 

Fixes for some of the various issues reported here:
https://github.com/dnbexperience/eufemia/issues/5528

I'll quote it here and update this list as I fix each one:

> - **BUG: Loops to selected:** 
> - Should loop to the last item when looping "up", not to the selected
item.

Fixed: no longer navigates to the item before or after the selected item
when looping.

> - **IMPROVEMENT: List not focused (Dropdown):** 
> - Should focus the list between loops when looping in both directions,
not just "up"

Fixed/Changed: now only focuses the ul when the list first opens. Once
user starts navigating only the items get focused.

> - **BUG/IMPROVEMENT: Doesn't loop (Autocomplete):**
>   - Should loop in both directions, not just "down"

Fixed: Autocomplete now loops in both directions

> - **IMPROVEMENT: Selected is not focused:** 
> - Should focus selected item when re-opening the list. (Focus priority
should be as follows: selected item > last focused item > the list.)

Implemented: if an item is selected, the focus is set to that item when
the list is re-opened.

> - **BUG: Search resets focus:** 
>   - Should only reset focus on search if there is a result.

Fixed: now search only changes focus if it finds a result to focus on

> - **IMPROVEMENT: we should probably not focus the ul list to be
accessibility compliant**
> -
https://www.w3.org/WAI/ARIA/apg/practices/keyboard-interface/#kbd_focus_activedescendant
states that when the controlling element receives focus the
aria-activedescendant should point to the item that is active. So there
should always be an active item. Though the idea is that the controlling
element always has DOM focus, and the items are just referenced and
visually styled. So we sohuld perhaps implement that properly.
> - Today we even fake it for screen readers by pretending the
activedecendant is the first or selected element. If not it can't start
navigating.

Fixed-ish: First, we make sure that there `aria-activedecendant` always
points to a valid item. Autocomplete was missing this. But we still
start with no item focused. In those cases the aria-activedescendant
defaults to the first item. This is a compromise between design and
accessibility, screen readers require an "active" item to be able to
properly navigate, but when you open a list we don't want to already
have an item focused unless it is the selected item.