feat(Badge): add variant `content` (#6079)
Builds on top of https://github.com/dnbexperience/eufemia/pull/6071

Deploy preview:
https://feat-badge-variant-content.eufemia-e25.pages.dev/uilib/components/badge/properties/

For now, I've not added any demos for the new variant content, as I'm
not sure if we should document it too much.
As of now it's mostly/only used internally. But since we'll add demos
for it in https://github.com/dnbexperience/eufemia/pull/6078, I think
it's correct to add the type to TS and our properties docs, but probably
nothing more 🙏